backend-python/requirements.txt
flask==3.0.2
flask-cors==4.0.0
boto3==1.34.69
python-dotenv==1.0.1 